@@ -18,26 +18,32 @@
             "name": "Stefan Kleff",
             "email": "s.kleff@goalio.de",
             "homepage": "http://www.goalio.de"
+        },
+        {
+            "name": "Bas Kamer",
+            "email": "baskamer@gmail.com",
+            "homepage": "https://bushbaby.nl"
         }
     ],
     "require": {
-        "php": "^5.5 || ^7.0",
-        "laminas/laminas-eventmanager": "^2.6 || ^3.0",
-        "laminas/laminas-servicemanager": "^2.7 || ^3.0",
-        "laminas/laminas-stdlib": "^2.2 || ^3.0",
-        "slm/queue": "^1.0",
-        "doctrine/dbal": "^2.5"
+        "php": "^7.2",
+        "laminas/laminas-eventmanager": "^3.2.1",
+        "laminas/laminas-servicemanager": "^3.3.1",
+        "laminas/laminas-stdlib": "^3.2",
+        "slm/queue": "^2.0",
+        "doctrine/dbal": "^2.10"
     },
     "require-dev": {
-        "laminas/laminas-config": "^2.2",
-        "laminas/laminas-modulemanager": "^2.5",
-        "laminas/laminas-view": "^2.5",
-        "laminas/laminas-log": "^2.5",
-        "laminas/laminas-i18n": "^2.5",
-        "laminas/laminas-serializer": "^2.5",
-        "laminas/laminas-mvc": "^2.5",
-        "phpunit/phpunit": "^4.8",
-        "squizlabs/php_codesniffer": "^2.7",
+        "laminas/laminas-config": "^2.6",
+        "laminas/laminas-modulemanager": "^2.8",
+        "laminas/laminas-view": "^2.10",
+        "laminas/laminas-log": "^2.9",
+        "laminas/laminas-i18n": "^2.6",
+        "laminas/laminas-serializer": "^2.8",
+        "laminas/laminas-mvc": "^2.7.13",
+        "phpunit/phpunit": "^8.5",
+        "doctrine/orm": "^2.6.3",
+        "squizlabs/php_codesniffer": "^3.5",
         "php-coveralls/php-coveralls": "^2.0"
     },
     "suggest": {
@@ -65,5 +71,9 @@
         "psr-4": {
             "SlmQueueDoctrineTest\\": "tests/"
         }
+    },
+    "scripts": {
+        "cs-check": "phpcs",
+        "cs-fix": "phpcbf"
     }
 }
