@@ -2,32 +2,37 @@
 
 namespace SlmQueueDoctrineTest\Factory;
 
-use PHPUnit_Framework_TestCase;
-use SlmQueue\Queue\QueuePluginManager;
+use PHPUnit\Framework\TestCase;
 use SlmQueueDoctrine\Factory\DoctrineQueueFactory;
 use SlmQueueDoctrine\Queue\DoctrineQueue;
 use SlmQueueDoctrineTest\Util\ServiceManagerFactory;
 
-class DoctrineQueueFactoryTest extends PHPUnit_Framework_TestCase
+class DoctrineQueueFactoryTest extends TestCase
 {
-    public function testCreateServiceGetsInstance()
+    public function testCreateServiceGetsInstance(): void
     {
-        $sm                 = ServiceManagerFactory::getServiceManager();
-        $factory            = new DoctrineQueueFactory();
-        $service            = $factory($sm, null);
+        $sm = ServiceManagerFactory::getServiceManager();
+        $factory = new DoctrineQueueFactory();
+        $service = $factory($sm, null);
 
         static::assertInstanceOf(DoctrineQueue::class, $service);
     }
 
-    public function testSpecifiedQueueOptionsOverrideModuleDefaults()
+    public function testSpecifiedQueueOptionsOverrideModuleDefaults(): void
     {
-        $sm                 = ServiceManagerFactory::getServiceManager();
-        $config             = $sm->get('config');
+        $sm = ServiceManagerFactory::getServiceManager();
+        $config = $sm->get('config');
 
-        $factory            = new DoctrineQueueFactory();
-        $service            = $factory($sm, 'my-doctrine-queue');
+        $factory = new DoctrineQueueFactory();
+        $service = $factory($sm, 'my-doctrine-queue');
 
-        static::assertEquals($service->getOptions()->getDeletedLifetime(), $config['slm_queue']['queues']['my-doctrine-queue']['deleted_lifetime']);
-        static::assertEquals($service->getOptions()->getBuriedLifetime(), $config['slm_queue']['queues']['my-doctrine-queue']['buried_lifetime']);
+        static::assertEquals(
+            $service->getOptions()->getDeletedLifetime(),
+            $config['slm_queue']['queues']['my-doctrine-queue']['deleted_lifetime']
+        );
+        static::assertEquals(
+            $service->getOptions()->getBuriedLifetime(),
+            $config['slm_queue']['queues']['my-doctrine-queue']['buried_lifetime']
+        );
     }
 }
