@@ -3,37 +3,38 @@
 namespace SlmQueueDoctrineTest\Listener\Strategy;
 
 use Doctrine\Common\Persistence\ObjectManager;
-use PHPUnit_Framework_TestCase;
+use Laminas\EventManager\EventManagerInterface;
+use PHPUnit\Framework\TestCase;
+use SlmQueue\Queue\QueueInterface;
+use SlmQueue\Strategy\AbstractStrategy;
 use SlmQueue\Worker\Event\AbstractWorkerEvent;
 use SlmQueue\Worker\Event\ProcessJobEvent;
 use SlmQueueDoctrine\Strategy\ClearObjectManagerStrategy;
 use SlmQueueDoctrine\Worker\DoctrineWorker;
 use SlmQueueDoctrineTest\Asset\OMJob;
-use SlmQueueTest\Asset\SimpleWorker;
-use Laminas\EventManager\EventManagerInterface;
 
-class ClearObjectManagerStrategyTest extends PHPUnit_Framework_TestCase
+class ClearObjectManagerStrategyTest extends TestCase
 {
     protected $queue;
     protected $worker;
     /** @var ClearObjectManagerStrategy */
     protected $listener;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->queue    = $this->getMock(\SlmQueue\Queue\QueueInterface::class);
-        $this->worker   = new DoctrineWorker($this->getMock(EventManagerInterface::class));
+        $this->queue = $this->createMock(QueueInterface::class);
+        $this->worker = new DoctrineWorker($this->createMock(EventManagerInterface::class));
         $this->listener = new ClearObjectManagerStrategy();
     }
 
-    public function testListenerInstanceOfAbstractStrategy()
+    public function testListenerInstanceOfAbstractStrategy(): void
     {
-        static::assertInstanceOf(\SlmQueue\Strategy\AbstractStrategy::class, $this->listener);
+        static::assertInstanceOf(AbstractStrategy::class, $this->listener);
     }
 
-    public function testListensToCorrectEventAtCorrectPriority()
+    public function testListensToCorrectEventAtCorrectPriority(): void
     {
-        $evm      = $this->getMock(EventManagerInterface::class);
+        $evm = $this->createMock(EventManagerInterface::class);
         $priority = 1;
 
         $evm->expects($this->at(0))->method('attach')
@@ -42,10 +43,10 @@ class ClearObjectManagerStrategyTest extends PHPUnit_Framework_TestCase
         $this->listener->attach($evm, $priority);
     }
 
-    public function testOnClearHandler()
+    public function testOnClearHandler(): void
     {
         $job = new OMJob();
-        $om  = $this->getMock(ObjectManager::class);
+        $om = $this->createMock(ObjectManager::class);
 
         $job->setObjectManager($om);
 
