@@ -2,36 +2,38 @@
 
 namespace SlmQueueDoctrineTest\Listener\Strategy;
 
-use PHPUnit_Framework_TestCase;
+use Laminas\EventManager\EventManagerInterface;
+use PHPUnit\Framework\TestCase;
+use SlmQueue\Queue\QueueInterface;
+use SlmQueue\Strategy\AbstractStrategy;
 use SlmQueue\Worker\Event\AbstractWorkerEvent;
 use SlmQueue\Worker\Event\ProcessIdleEvent;
 use SlmQueueDoctrine\Queue\DoctrineQueueInterface;
 use SlmQueueDoctrine\Strategy\IdleNapStrategy;
 use SlmQueueDoctrine\Worker\DoctrineWorker;
-use Laminas\EventManager\EventManagerInterface;
 
-class IdleNapStrategyTest extends PHPUnit_Framework_TestCase
+class IdleNapStrategyTest extends TestCase
 {
     protected $queue;
     protected $worker;
     /** @var IdleNapStrategy */
     protected $listener;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->queue    = $this->getMock(\SlmQueue\Queue\QueueInterface::class);
-        $this->worker   = new DoctrineWorker($this->getMock(EventManagerInterface::class));
+        $this->queue = $this->createMock(QueueInterface::class);
+        $this->worker = new DoctrineWorker($this->createMock(EventManagerInterface::class));
         $this->listener = new IdleNapStrategy();
     }
 
-    public function testListenerInstanceOfAbstractStrategy()
+    public function testListenerInstanceOfAbstractStrategy(): void
     {
-        static::assertInstanceOf(\SlmQueue\Strategy\AbstractStrategy::class, $this->listener);
+        static::assertInstanceOf(AbstractStrategy::class, $this->listener);
     }
 
-    public function testListensToCorrectEventAtCorrectPriority()
+    public function testListensToCorrectEventAtCorrectPriority(): void
     {
-        $evm      = $this->getMock(EventManagerInterface::class);
+        $evm = $this->createMock(EventManagerInterface::class);
         $priority = 1;
 
         $evm->expects($this->at(0))->method('attach')
@@ -40,21 +42,21 @@ class IdleNapStrategyTest extends PHPUnit_Framework_TestCase
         $this->listener->attach($evm, $priority);
     }
 
-    public function testNapDurationDefault()
+    public function testNapDurationDefault(): void
     {
         static::assertEquals(1, $this->listener->getNapDuration());
     }
 
-    public function testNapDurationSetter()
+    public function testNapDurationSetter(): void
     {
         $this->listener->setNapDuration(2);
 
         static::assertEquals(2, $this->listener->getNapDuration());
     }
 
-    public function testOnIdleHandler()
+    public function testOnIdleHandler(): void
     {
-        $this->queue = $this->getMock(DoctrineQueueInterface::class);
+        $this->queue = $this->createMock(DoctrineQueueInterface::class);
 
         $start_time = microtime(true);
         $this->listener->onIdle(new ProcessIdleEvent($this->worker, $this->queue));
@@ -62,7 +64,7 @@ class IdleNapStrategyTest extends PHPUnit_Framework_TestCase
         static::assertGreaterThan(1, $elapsed_time);
 
 
-        $this->queue    = $this->getMock(\SlmQueue\Queue\QueueInterface::class);
+        $this->queue = $this->createMock(QueueInterface::class);
 
         $start_time = microtime(true);
         $this->listener->onIdle(new ProcessIdleEvent($this->worker, $this->queue));
