@@ -29,10 +29,8 @@ class BuryableException extends RuntimeException implements ExceptionInterface
 
     /**
      * Get the options
-     *
-     * @return array
      */
-    public function getOptions()
+    public function getOptions(): array
     {
         return $this->options;
     }
