@@ -14,30 +14,20 @@ interface DoctrineQueueInterface extends QueueInterface
      * @param  array        $options
      * @return void
      */
-    public function release(JobInterface $job, array $options = []);
+    public function release(JobInterface $job, array $options = []): void;
 
     /**
      * Bury a job. When a job is buried, it won't be retrieved from the queue
-     *
-     * @param  JobInterface $job
-     * @param  array        $options
-     * @return void
      */
-    public function bury(JobInterface $job, array $options = []);
+    public function bury(JobInterface $job, array $options = []): void;
 
     /**
      * Recover jobs which are in the state 'running' for more then $executionTime minutes
-     *
-     * @param  int   $executionTime
-     * @return mixed
      */
-    public function recover($executionTime);
+    public function recover(int $executionTime): int;
 
     /**
      * Get a job from the queue without processing it
-     *
-     * @param  int          $id
-     * @return JobInterface
      */
-    public function peek($id);
+    public function peek(int $id): JobInterface;
 }
