@@ -2,19 +2,19 @@
 
 namespace SlmQueueDoctrine\Strategy;
 
-use DoctrineModule\Persistence\ObjectManagerAwareInterface as DoctrineModuleObjectManagerAwareInterface;
-use SlmQueueDoctrine\Persistence\ObjectManagerAwareInterface;
+use DoctrineModule\Persistence\ObjectManagerAwareInterface as DMObjectManagerAwareInterface;
+use Laminas\EventManager\EventManagerInterface;
 use SlmQueue\Strategy\AbstractStrategy;
 use SlmQueue\Worker\Event\AbstractWorkerEvent;
 use SlmQueue\Worker\Event\ProcessJobEvent;
-use Laminas\EventManager\EventManagerInterface;
+use SlmQueueDoctrine\Persistence\ObjectManagerAwareInterface;
 
 class ClearObjectManagerStrategy extends AbstractStrategy
 {
     /**
      * {@inheritDoc}
      */
-    public function attach(EventManagerInterface $events, $priority = 1)
+    public function attach(EventManagerInterface $events, $priority = 1): void
     {
         $this->listeners[] = $events->attach(
             AbstractWorkerEvent::EVENT_PROCESS_JOB,
@@ -23,18 +23,20 @@ class ClearObjectManagerStrategy extends AbstractStrategy
         );
     }
 
-    /**
-     * @param ProcessJobEvent $event
-     */
-    public function onClear(ProcessJobEvent $event)
+    public function onClear(ProcessJobEvent $event): void
     {
         /** @var ObjectManagerAwareInterface $job */
         $job = $event->getJob();
 
-        if (($job instanceof ObjectManagerAwareInterface || $job instanceof DoctrineModuleObjectManagerAwareInterface)
-            && $job->getObjectManager()
-        ) {
-            $job->getObjectManager()->clear();
+
+        if (! ($job instanceof ObjectManagerAwareInterface || $job instanceof DMObjectManagerAwareInterface)) {
+            return;
         }
+
+        if (!$manager = $job->getObjectManager()) {
+            return;
+        }
+
+        $manager->clear();
     }
 }
