@@ -17,14 +17,15 @@ class OMJob extends AbstractJob implements ObjectManagerAwareInterface
     /**
      * {@inheritDoc}
      */
-    public function execute()
+    public function execute(): ?int
     {
+        return null;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function setObjectManager(ObjectManager $objectManager)
+    public function setObjectManager(ObjectManager $objectManager): void
     {
         $this->objectManager = $objectManager;
     }
@@ -32,7 +33,7 @@ class OMJob extends AbstractJob implements ObjectManagerAwareInterface
     /**
      * {@inheritDoc}
      */
-    public function getObjectManager()
+    public function getObjectManager(): ObjectManager
     {
         return $this->objectManager;
     }
