@@ -9,7 +9,8 @@ class SimpleJob extends AbstractJob
     /**
      * {@inheritDoc}
      */
-    public function execute()
+    public function execute(): ?int
     {
+        return null;
     }
 }
