@@ -2,25 +2,26 @@
 
 namespace SlmQueueDoctrineTest;
 
+use PHPUnit\Framework\TestCase as TestCase;
+use SlmQueueDoctrine\ConfigProvider;
 use SlmQueueDoctrine\Module;
-use PHPUnit_Framework_TestCase as TestCase;
 
 class ConfigProviderTest extends TestCase
 {
-    public function testConfigProviderGetConfig()
+    public function testConfigProviderGetConfig(): void
     {
-        $configProvider = new \SlmQueueDoctrine\ConfigProvider();
-        $config         = $configProvider();
+        $configProvider = new ConfigProvider();
+        $config = $configProvider();
 
         static::assertNotEmpty($config);
     }
 
-    public function testConfigEqualsToModuleConfig()
+    public function testConfigEqualsToModuleConfig(): void
     {
-        $module         = new Module();
-        $moduleConfig   = $module->getConfig();
-        $configProvider = new \SlmQueueDoctrine\ConfigProvider();
-        $config         = $configProvider();
+        $module = new Module();
+        $moduleConfig = $module->getConfig();
+        $configProvider = new ConfigProvider();
+        $config = $configProvider();
 
         static::assertEquals($moduleConfig['service_manager'], $config['dependencies']);
         static::assertEquals($moduleConfig['slm_queue'], $config['slm_queue']);
