@@ -2,12 +2,12 @@
 
 namespace SlmQueueDoctrineTest\Controller;
 
-use PHPUnit_Framework_TestCase as TestCase;
-use SlmQueueDoctrine\Controller\DoctrineWorkerController;
-use SlmQueueDoctrineTest\Util\ServiceManagerFactory;
 use Laminas\Mvc\Router\RouteMatch;
 use Laminas\ServiceManager\Exception\ServiceNotFoundException;
 use Laminas\ServiceManager\ServiceManager;
+use PHPUnit\Framework\TestCase;
+use SlmQueueDoctrine\Controller\DoctrineWorkerController;
+use SlmQueueDoctrineTest\Util\ServiceManagerFactory;
 
 class DoctrineWorkerControllerTest extends TestCase
 {
@@ -16,19 +16,20 @@ class DoctrineWorkerControllerTest extends TestCase
      */
     protected $serviceManager;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
         $this->serviceManager = ServiceManagerFactory::getServiceManager();
     }
 
-    public function testThrowExceptionIfQueueIsUnknown()
+    public function testThrowExceptionIfQueueIsUnknown(): void
     {
         $controller = $this->serviceManager->get('ControllerLoader')->get(DoctrineWorkerController::class);
         $routeMatch = new RouteMatch(['queue' => 'unknownQueue']);
         $controller->getEvent()->setRouteMatch($routeMatch);
 
-        $this->setExpectedException(ServiceNotFoundException::class);
+        $this->expectException(ServiceNotFoundException::class);
+
         $controller->processAction();
     }
 }
